refactor(#10): source files were moved to the src directory
