+++ backend/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "style": "eslint .",
     "style:fix": "npm run style -- --fix",
-    "start": "node ./bin/www",
+    "start": "node src/bin/www",
     "start:dev": "nodemon start",
     "test": "jest",
     "test:coverage" : "npm run test -- --coverage",
